♻ Quality gate refactoring
Removed some double code and made into reusable components/styles instead. Also modified a regex to prevent DOS.
