+++ app/Domits/src/features/hostdashboard/hosthelpdesk/screens/HostHelpDesk.js
@@ -25,7 +25,7 @@ export default function Contact() {
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [feedbackMessage, setFeedbackMessage] = useState('');
 
-  const isValidEmail = email => /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email);
+  const isValidEmail = email => /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/.test(email);
 
   // const pickAttachments = async () => {
   //   try {
