+++ app/Domits/src/screens/profile/utils/ProfileFunctions.js
@@ -1,0 +1,40 @@
+import {Alert} from "react-native";
+import {deleteUser} from "../../GeneralUtils/GenUtils";
+import NavigateTo from "../../../navigation/NavigationFunctions";
+
+export const setUserAttributes = async (userAttributes, setFirstName, setEmailAddress) => {
+    try {
+        setFirstName(userAttributes?.given_name);
+        setEmailAddress(userAttributes?.email);
+    } catch (error) {
+        console.error('Error fetching user attributes:', error);
+    }
+};
+
+export const deleteAccount = async (navigation, userId) => {
+    Alert.alert(
+        'Delete Account',
+        'Are you sure you want to delete your account? This action cannot be undone.',
+        [
+            {
+                text: 'Cancel',
+                onPress: () => console.log('Cancel Pressed'),
+                style: 'cancel',
+            },
+            {
+                text: 'Delete',
+                onPress: async () => {
+                    try {
+                        await deleteUser(userId); // Ensure account deletion completes
+                        NavigateTo(navigation).login();
+                    } catch (error) {
+                        console.error('Failed to delete account:', error);
+                        alert('Error deleting account. Please try again.');
+                    }
+                },
+                style: 'destructive', // Makes the button appear as a destructive action on iOS
+            },
+        ],
+        {cancelable: true}, // Allows tapping outside to cancel
+    );
+};
